Server/requirements.txt
fastapi
uvicorn
python-dotenv
langchain_chroma
langchain_huggingface
sentence_transformers
langchain_community
langchain
langchain-google-genai
PyPDF2
pydantic
pymongo[srv]==3.12
pydantic-settings
PyJWT
passlib